# Represent every values as text